requirements.txt
# Core dependencies
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.4.0

# Time series models
statsmodels==0.14.0
pmdarima==2.0.3
prophet==1.1.5
cmdstanpy==1.2.0

# Machine learning ops
mlflow==2.9.2
optuna==3.4.0

# API and web
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Data validation
pydantic-settings==2.1.0
pandera==0.17.2

# Visualization
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.17.0

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
joblib==1.3.2
scipy==1.11.3

# Database (optional)
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Monitoring and logging
prometheus-client==0.19.0
loguru==0.7.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.2

# Development
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1
pre-commit==3.5.0

# Deployment
gunicorn==21.2.0
docker==6.1.3

